♻️ refactor: separate type in Radio component
