+++ packages/radio/src/types/base.ts
@@ -1,0 +1,2 @@
+export type RadioSize = 'sm' | 'md' | 'lg';
+export type RadioValueType = string | number | boolean;
