+++ packages/radio/src/types/props.ts
@@ -1,17 +1,5 @@
+import { RadioSize, RadioValueType } from './base';
 import type { ColorToken } from '@jdesignlab/theme';
-import type { EventType } from '@jdesignlab/utils';
-import type { Dispatch, SetStateAction } from 'react';
-
-export type RadioSize = 'sm' | 'md' | 'lg';
-export type RadioValueType = string | number | boolean;
-export type RadioAttributes = { [key: string]: string | ((event: EventType) => void) };
-export interface ReturnContext {
-  defaultValue?: string;
-  name: string;
-  rootProps: RadioAttributes | null;
-  setValue: Dispatch<SetStateAction<string>> | null;
-  value: RadioValueType | null;
-}
 
 export interface RadioLabelProps {
   id: string;
